Move projects section to separate component
