+++ src/pages/HomePage.tsx
@@ -1,7 +1,6 @@
 import tsu_logo from "images/tech-start-logo-white.png";
 import "./HomePage.css";
 import { motion } from "framer-motion/dist/framer-motion";
-import Lottie from "react-lottie";
 
 import NumberStat from "components/NumberStat";
 import Divider from "components/Divider";
@@ -9,21 +8,10 @@ import Particle from "components/Particles";
 import SponsorSection from "components/SponsorSection/SponsorSection";
 import Blobbie from "components/Blobbie";
 import EventsSection from "components/EventsSection/EventsSection";
-import rocket from "../images/lottie/rocket.json";
+import { ProjectsSection } from "./ProjectsSection";
 import ApplyButton from "components/ApplyButton";
-import HoverButton from "components/HoverButton/HoverButton";
-import { ButtonMode } from "components/HoverButton/HoverButton.styles";
 
 const HomePage = () => {
-  const rocketLottieOptions = {
-    loop: true,
-    autoplay: true,
-    animationData: rocket,
-    rendererSettings: {
-      preserveAspectRatio: "xMidYMid slice",
-    },
-  };
-
   return (
     <div className="homePage">
       <header className="homePage__hero" id="homePageTop">
@@ -106,46 +94,6 @@ const HomePage = () => {
           </div>
         </div>
       </section>
-      <section className="homePage__section--projects">
-        <div className="homePage__container">
-          <h1 className="chonkyHeading chonkyHeading--white chonkyHeading--lessMargin">
-            TechStart Projects
-          </h1>
-          <Divider />
-          <div className="homepage__section--projects-description-div">
-            <div style={{ marginLeft: "10vw" }}>
-              <p className="regularText regularText--white" id="draft">
-                Our members form teams that gather together to brainstorm a
-                project, and then work on that idea collaboratively with 6~9
-                other creators over the year. During this time, the projects
-                evolve and shift, and our members get to experience the full
-                lifecycle of bringing a complex idea to life. Each project team
-                consists of software developers, business strategist, project
-                designer, and a project manager.
-              </p>
-              <div className="projectButtonDiv">
-                <HoverButton
-                  mode={ButtonMode.GRADIENT}
-                  glowOnHover={true}
-                  link={"/projects"}
-                  text={"Check out our projects!"}
-                  linkIsInternal={true}
-                />
-              </div>
-            </div>
-            <div />
-          </div>
-          <div className="homepage__section--projects-div">
-            <div />
-            <div style={{ width: "25vw", height: "25vw" }}>
-              <Lottie
-                options={rocketLottieOptions}
-                style={{ marginLeft: "15vw" }}
-              />
-            </div>
-          </div>
-        </div>
-      </section>
       <Blobbie
         id={2}
         width={100}
@@ -172,6 +120,8 @@ const HomePage = () => {
           <EventsSection />
         </div>
       </section>
+
+      <ProjectsSection />
 
       <section className="homePage__section homePage__section--sponsors">
         <Blobbie
